e-commerce/requirements.txt
pymongo==4.5
pydantic==1.10.2
requests==2.31
email-validator==2.0
typing-extensions==4.8
django==4.2
django-allauth==0.43.0
django-ninja-extra==0.19.8

django-cors-headers==3.8.0

gunicorn==21.2